srcs/django/auth_system/authentication.py
from rest_framework_simplejwt.authentication import JWTAuthentication
from rest_framework.exceptions import AuthenticationFailed

class CookieJWTAuthentication(JWTAuthentication):
	def authenticate(self, request):
		raw_token = request.COOKIES.get('access_token')
		if raw_token is None:
			return None
		try:
			validated_token = self.get_validated_token(raw_token)
		except Exception as e:
			raise AuthenticationFailed(f'Error validating token: {str(e)}')
		
		return self.get_user(validated_token), validated_token
